requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
reportlab==4.0.7
gunicorn==21.2.0
